+++ DEV_ONLY/App.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import {render} from 'react-dom';
 
-import hashIt from '../src';
-import * as utils from '../src/utils';
+import hash from '../src';
 
 document.body.style.backgroundColor = '#1d1d1d';
 document.body.style.color = '#d5d5d5';
@@ -99,22 +98,22 @@ const object = {
 const profile = (iterations = 100) => {
   let index = -1;
 
-  console.profile('hashIt timing');
+  console.profile('hash timing');
 
   while (++index < iterations) {
-    hashIt(object);
-    hashIt(a);
+    hash(object);
+    hash(a);
 
     for (let key in object) {
-      hashIt(object[key]);
+      hash(object[key]);
     }
 
-    hashIt(document.body);
+    hash(document.body);
     // this is the killer, so only profiled if you uncomment
-    // hashIt(window);
+    // hash(window);
   }
 
-  console.profileEnd('hashIt timing');
+  console.profileEnd('hash timing');
 
   console.log('Check the Profiles tab in DevTools to see the output.');
 };
@@ -124,104 +123,106 @@ const benchmark = () => {
 };
 
 const visualValidation = (iterations = 100) => {
-  console.log(object, hashIt(object));
-  console.log(a, hashIt(a));
-  console.log(object.string, hashIt(object.string));
-  console.log(object.num, hashIt(object.num));
-  console.log(object.bool, hashIt(object.bool));
-  console.log(object.func, hashIt(object.func));
-  console.log(object.undef, hashIt(object.undef));
-  console.log(object.nil, hashIt(object.nil));
-  console.log(object.obj, hashIt(object.obj));
-  console.log(object.arr, hashIt(object.arr));
-  console.log(object.el, hashIt(object.el));
-  console.log(object.fragment, hashIt(object.fragment));
-  console.log(object.svg, hashIt(object.svg));
-  console.log(object.regexp, hashIt(object.regexp));
-  console.log(object.event, hashIt(object.event));
-  console.log(object.nodeList, hashIt(object.nodeList));
+  console.log(object, hash(object));
+  console.log(a, hash(a));
+  console.log(object.string, hash(object.string));
+  console.log(object.num, hash(object.num));
+  console.log(object.bool, hash(object.bool));
+  console.log(object.func, hash(object.func));
+  console.log(object.undef, hash(object.undef));
+  console.log(object.nil, hash(object.nil));
+  console.log(object.obj, hash(object.obj));
+  console.log(object.arr, hash(object.arr));
+  console.log(object.el, hash(object.el));
+  console.log(object.fragment, hash(object.fragment));
+  console.log(object.svg, hash(object.svg));
+  console.log(object.regexp, hash(object.regexp));
+  console.log(object.event, hash(object.event));
+  console.log(object.nodeList, hash(object.nodeList));
 
   // comment out for older browser testing
-  console.log(object.symbol, hashIt(object.symbol));
-  console.log(object.err, hashIt(object.err));
-  console.log(object.map, hashIt(object.map));
-  console.log(object.set, hashIt(object.set));
-  console.log(object.weakMap, hashIt(object.weakMap));
-  console.log(object.weakSet, hashIt(object.weakSet));
-  console.log(object.dataView, hashIt(object.dataView));
-  console.log(object.arrayBuffer, hashIt(object.arrayBuffer));
-  console.log(object.float32Array, hashIt(object.float32Array));
-  console.log(object.float64Array, hashIt(object.float64Array));
-  console.log(object.generator, hashIt(object.generator));
-  console.log(object.int8Array, hashIt(object.int8Array));
-  console.log(object.int16Array, hashIt(object.int16Array));
-  console.log(object.int32Array, hashIt(object.int32Array));
-  console.log(object.promise, hashIt(object.promise));
-  console.log(object.uint8Array, hashIt(object.uint8Array));
-  console.log(object.uint8ClampedArray, hashIt(object.uint8ClampedArray));
-  console.log(object.uint16Array, hashIt(object.uint16Array));
-  console.log(object.uint32Array, hashIt(object.uint32Array));
+  console.log(object.symbol, hash(object.symbol));
+  console.log(object.err, hash(object.err));
+  console.log(object.map, hash(object.map));
+  console.log(object.set, hash(object.set));
+  console.log(object.weakMap, hash(object.weakMap));
+  console.log(object.weakSet, hash(object.weakSet));
+  console.log(object.dataView, hash(object.dataView));
+  console.log(object.arrayBuffer, hash(object.arrayBuffer));
+  console.log(object.float32Array, hash(object.float32Array));
+  console.log(object.float64Array, hash(object.float64Array));
+  console.log(object.generator, hash(object.generator));
+  console.log(object.int8Array, hash(object.int8Array));
+  console.log(object.int16Array, hash(object.int16Array));
+  console.log(object.int32Array, hash(object.int32Array));
+  console.log(object.promise, hash(object.promise));
+  console.log(object.uint8Array, hash(object.uint8Array));
+  console.log(object.uint8ClampedArray, hash(object.uint8ClampedArray));
+  console.log(object.uint16Array, hash(object.uint16Array));
+  console.log(object.uint32Array, hash(object.uint32Array));
 
-  console.log(object.ReactStatefulClass, hashIt(object.ReactStatefulClass));
-  console.log(object.ReactStatefulElement, hashIt(object.ReactStatefulElement));
-  console.log(object.ReactStatelessClass, hashIt(object.ReactStatelessClass));
-  console.log(object.ReactStatelessElement, hashIt(object.ReactStatelessElement));
-  console.log(object.doc.body, hashIt(object.doc.body));
-  console.log(object.win, hashIt(object.win));
+  console.log(object.ReactStatefulClass, hash(object.ReactStatefulClass));
+  console.log(object.ReactStatefulElement, hash(object.ReactStatefulElement));
+  console.log(object.ReactStatelessClass, hash(object.ReactStatelessClass));
+  console.log(object.ReactStatelessElement, hash(object.ReactStatelessElement));
+  console.log(object.doc.body, hash(object.doc.body));
+  console.log(object.win, hash(object.win));
 };
 
 const hashOnlyValidation = (iterations = 100) => {
-  console.log(hashIt(object));
-  console.log(hashIt(a));
-  console.log(hashIt(object.string));
-  console.log(hashIt(object.num));
-  console.log(hashIt(object.bool));
-  console.log(hashIt(object.func));
-  console.log(hashIt(object.undef));
-  console.log(hashIt(object.nil));
-  console.log(hashIt(object.obj));
-  console.log(hashIt(object.arr));
-  console.log(hashIt(object.el));
-  console.log(hashIt(object.fragment));
-  console.log(hashIt(object.svg));
-  console.log(hashIt(object.math));
-  console.log(hashIt(object.regexp));
-  console.log(hashIt(object.event));
-  console.log(hashIt(object.nodeList));
+  console.log(hash(object));
+  console.log(hash(a));
+  console.log(hash(object.string));
+  console.log(hash(object.num));
+  console.log(hash(object.bool));
+  console.log(hash(object.func));
+  console.log(hash(object.undef));
+  console.log(hash(object.nil));
+  console.log(hash(object.obj));
+  console.log(hash(object.arr));
+  console.log(hash(object.el));
+  console.log(hash(object.fragment));
+  console.log(hash(object.svg));
+  console.log(hash(object.math));
+  console.log(hash(object.regexp));
+  console.log(hash(object.event));
+  console.log(hash(object.nodeList));
 
   // comment out for older browser testing
-  console.log(hashIt(object.symbol));
-  console.log(hashIt(object.err));
-  console.log(hashIt(object.map));
-  console.log(hashIt(object.set));
-  console.log(hashIt(object.weakMap));
-  console.log(hashIt(object.weakSet));
-  console.log(hashIt(object.arrayBuffer));
-  console.log(hashIt(object.dataView));
-  console.log(hashIt(object.float32Array));
-  console.log(hashIt(object.float64Array));
-  console.log(hashIt(object.generator));
-  console.log(hashIt(object.int8Array));
-  console.log(hashIt(object.int16Array));
-  console.log(hashIt(object.int32Array));
-  console.log(hashIt(object.promise));
-  console.log(hashIt(object.uint8Array));
-  console.log(hashIt(object.uint8ClampedArray));
-  console.log(hashIt(object.uint16Array));
-  console.log(hashIt(object.uint32Array));
+  console.log(hash(object.symbol));
+  console.log(hash(object.err));
+  console.log(hash(object.map));
+  console.log(hash(object.set));
+  console.log(hash(object.weakMap));
+  console.log(hash(object.weakSet));
+  console.log(hash(object.arrayBuffer));
+  console.log(hash(object.dataView));
+  console.log(hash(object.float32Array));
+  console.log(hash(object.float64Array));
+  console.log(hash(object.generator));
+  console.log(hash(object.int8Array));
+  console.log(hash(object.int16Array));
+  console.log(hash(object.int32Array));
+  console.log(hash(object.promise));
+  console.log(hash(object.uint8Array));
+  console.log(hash(object.uint8ClampedArray));
+  console.log(hash(object.uint16Array));
+  console.log(hash(object.uint32Array));
 
-  console.log(hashIt(object.ReactStatefulClass));
-  console.log(hashIt(object.ReactStatefulElement));
-  console.log(hashIt(object.ReactStatelessClass));
-  console.log(hashIt(object.ReactStatelessElement));
-  console.log(hashIt(document.body));
-  console.log(hashIt(window));
+  console.log(hash(object.ReactStatefulClass));
+  console.log(hash(object.ReactStatefulElement));
+  console.log(hash(object.ReactStatelessClass));
+  console.log(hash(object.ReactStatelessElement));
+  console.log(hash(document.body));
+  console.log(hash(window));
 };
 
 // benchmark();
 // profile(1000);
 visualValidation();
 // hashOnlyValidation();
+
+console.log(hash({foo: 'bar'}));
 
 const div = document.createElement('div');